feat: create updateCustomerUseCase

+++ src/@core/adapters/typeorm/customer-typeorm.repository.ts
@@ -12,13 +12,43 @@ export class CustomerTypeOrmRepository implements CustomerRepositoryInterface {
     return this.ormRepository.find();
   }
 
-  findById(customerId: string): Promise<Customer> {
-    return this.ormRepository.findOneBy({
-      id: customerId,
+  async findById(id: string): Promise<Customer> {
+    const customerExists = await this.ormRepository.findOneBy({
+      id: id,
+    });
+    if (!customerExists) {
+      throw new Error('Customer is not exist');
+    }
+    return await this.ormRepository.findOneBy({
+      id: id,
     });
   }
-  delete(id: string): Promise<void> {
-    this.ormRepository.delete({ id });
+
+  async update(id: string, customer: Customer): Promise<Customer | void> {
+    const customerExists = await this.ormRepository.findOneBy({
+      id: id,
+    });
+
+    if (!customerExists) {
+      throw new Error('Customer is not exist');
+    }
+    this.ormRepository.update(
+      { id: id },
+      {
+        cnpj: customer.cnpj,
+        consultantsTotal: customer.consultantsTotal,
+        fantasyName: customer.fantasyName,
+      },
+    );
     return;
+  }
+  async delete(id: string) {
+    const customerExists = this.ormRepository.findOne({ where: { id } });
+    console.log(customerExists);
+    if (!customerExists) {
+      console.log('não existe');
+      return new Error('Customer is not exist');
+    }
+    return this.ormRepository.delete({ id });
   }
 }
